+++ package.json
@@ -26,7 +26,7 @@
     "registry": "https://registry.npmjs.org/",
     "access": "public"
   },
-  "main": "index.js",
+  "main": "src/index.js",
   "bin": {
     "ipfs-deploy": "bin/ipfs-deploy.js",
     "ipd": "bin/ipfs-deploy.js"
